requirements.txt
Django == 1.9.8
MySQL - python == 1.2.5
djangorestframework == 3.4.0
geopy == 1.11
requests == 2.10
gunicorn
dj - database - url == 0.3.0
whitenoise
